Workspaces: extract WorkspaceCore — the overlay/cache semantics as a host-agnostic library object
The DO is now a thin host (cloudflare/workspace's constructor shape), so the same core can later be hosted by the Agent DO or Repo DO, and the storage engine stays swappable.
